sn32: 2xx: GPT: cleanup the configuration
add missing hooks for GPTD2
fixup the prescaler assert check
update the interrupt handler

+++ os/hal/ports/SN32/LLD/SN32F2xx/CT/hal_gpt_lld.h
@@ -139,7 +139,7 @@ typedef uint32_t gptfreq_t;
 /**
  * @brief   GPT counter type.
  */
-typedef uint16_t gptcnt_t;
+typedef uint32_t gptcnt_t;
 
 /**
  * @brief   Driver configuration structure.
